requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-socketio>=5.3.0
pystray>=0.19.5
pillow>=10.0.0
psutil>=5.9.0
pyinstaller>=6.0.0
